requirements.txt
fastapi[all]>=0.111.0
motor>=3.4.0
beanie>=1.26.0
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
pydantic>=2.7.1
google-generativeai>=0.5.4
websockets>=12.0
uvicorn>=0.29.0
httpx>=0.27.0
pydub>=0.25.1
twilio>=9.0.4
jinja2>=3.1.4
